Spawn nested modules dynamically, Add IInterfaceConnector, Move core classes to Core namespace
