+++ Scripts/Core/EditorVR.UI.cs
@@ -6,21 +6,82 @@ using UnityEditor.Experimental.EditorVR.Utilities;
 using UnityEngine;
 using UnityEngine.EventSystems;
 
-namespace UnityEditor.Experimental.EditorVR
+namespace UnityEditor.Experimental.EditorVR.Core
 {
 	partial class EditorVR
 	{
 		[SerializeField]
 		Camera m_EventCameraPrefab;
 
-		class UI : Nested
+		class UI : Nested, IInterfaceConnector
 		{
-			internal List<IManipulatorVisibility> manipulatorVisibilities { get { return m_ManipulatorVisibilities; } }
-			readonly List<IManipulatorVisibility> m_ManipulatorVisibilities = new List<IManipulatorVisibility>();
+			const byte k_MinStencilRef = 2;
 
+			byte stencilRef
+			{
+				get { return m_StencilRef; }
+				set
+				{
+					m_StencilRef = (byte)Mathf.Clamp(value, k_MinStencilRef, byte.MaxValue);
+
+					// Wrap
+					if (m_StencilRef == byte.MaxValue)
+						m_StencilRef = k_MinStencilRef;
+				}
+			}
+			byte m_StencilRef = k_MinStencilRef;
+
+			readonly List<IManipulatorVisibility> m_ManipulatorVisibilities = new List<IManipulatorVisibility>();
 			readonly HashSet<ISetManipulatorsVisible> m_ManipulatorsHiddenRequests = new HashSet<ISetManipulatorsVisible>();
 
 			internal Camera eventCamera { get; private set; }
+
+			public void ConnectInterface(object obj, Transform rayOrigin = null)
+			{
+				var instantiateUI = obj as IInstantiateUI;
+				if (instantiateUI != null)
+					instantiateUI.instantiateUI = InstantiateUI;
+
+				var manipulatorVisiblity = obj as IManipulatorVisibility;
+				if (manipulatorVisiblity != null)
+					m_ManipulatorVisibilities.Add(manipulatorVisiblity);
+
+				var setManipulatorsVisible = obj as ISetManipulatorsVisible;
+				if (setManipulatorsVisible != null)
+					setManipulatorsVisible.setManipulatorsVisible = SetManipulatorsVisible;
+
+				var usesStencilRef = obj as IUsesStencilRef;
+				if (usesStencilRef != null)
+				{
+					byte? stencilRef = null;
+
+					var mb = obj as MonoBehaviour;
+					if (mb)
+					{
+						var parent = mb.transform.parent;
+						if (parent)
+						{
+							// For workspaces and tools, it's likely that the stencil ref should be shared internally
+							var parentStencilRef = parent.GetComponentInParent<IUsesStencilRef>();
+							if (parentStencilRef != null)
+								stencilRef = parentStencilRef.stencilRef;
+						}
+					}
+
+					usesStencilRef.stencilRef = stencilRef ?? RequestStencilRef();
+				}
+
+				var requestStencilRef = obj as IRequestStencilRef;
+				if (requestStencilRef != null)
+					requestStencilRef.requestStencilRef = RequestStencilRef;
+			}
+
+			public void DisconnectInterface(object obj)
+			{
+				var manipulatorVisiblity = obj as IManipulatorVisibility;
+				if (manipulatorVisiblity != null)
+					m_ManipulatorVisibilities.Remove(manipulatorVisiblity);
+			}
 
 			internal void Initialize()
 			{
@@ -63,7 +124,7 @@ namespace UnityEditor.Experimental.EditorVR
 				return go;
 			}
 
-			internal void SetManipulatorsVisible(ISetManipulatorsVisible setter, bool visible)
+			void SetManipulatorsVisible(ISetManipulatorsVisible setter, bool visible)
 			{
 				if (visible)
 					m_ManipulatorsHiddenRequests.Remove(setter);
@@ -76,6 +137,11 @@ namespace UnityEditor.Experimental.EditorVR
 				var manipulatorsVisible = m_ManipulatorsHiddenRequests.Count == 0;
 				foreach (var mv in m_ManipulatorVisibilities)
 					mv.manipulatorVisible = manipulatorsVisible;
+			}
+
+			byte RequestStencilRef()
+			{
+				return stencilRef++;
 			}
 		}
 	}
