+++ Scripts/Core/InterfaceConnectors.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a27492dadb8910149a9458ac95347715
+folderAsset: yes
+timeCreated: 1489520314
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
