+++ Scripts/Core/InterfaceConnectors/HierarchyModuleConnector.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5853c284ea4ca5d4492cbb8506758cc0
+timeCreated: 1489520545
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
