+++ Scripts/Core/InterfaceConnectors/HighlightModuleConnector.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 39b5983bf5da6da41a2a512df83ccb33
+timeCreated: 1489520545
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
