+++ Scripts/Core/InterfaceConnectors/LockModuleConnector.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cd6ca3d67dbdb944b9613c75d5b6a279
+timeCreated: 1489520545
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
