+++ Scripts/Core/InterfaceConnectors/SelectionModuleConnector.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 79c92ee0c472e1149b70a7239e5154f4
+timeCreated: 1489520545
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
