--- Scripts/Core/Interfaces/IBinding.cs
@@ -1,12 +1,0 @@
-﻿using UnityEngine;
-
-namespace UnityEditor.Experimental.EditorVR.Core
-{
-	delegate void BindDelegate();
-	
-	public interface IBinding<T> where T : class
-	{
-		void Bind();
-		void ConnectInterface(T obj, Transform rayOrigin = null);
-	}
-}
