+++ Scripts/Core/Interfaces/IInterfaceConnector.cs.meta
