+++ Scripts/Helpers/VRSmoothCamera.cs
@@ -1,5 +1,6 @@
 ﻿#if UNITY_EDITOR && UNITY_EDITORVR
 using System;
+using UnityEditor.Experimental.EditorVR.Core;
 using UnityEngine;
 
 namespace UnityEditor.Experimental.EditorVR.Helpers
