+++ Scripts/Interfaces/IUsesRayLocking.cs
@@ -9,19 +9,37 @@ namespace UnityEditor.Experimental.EditorVR
 	/// </summary>
 	public interface IUsesRayLocking
 	{
+	}
+
+	public static class IUsesRayLockingMethods
+	{
+		internal delegate bool RayLockingDelegate(Transform rayOrigin, object obj);
+
+		internal static RayLockingDelegate lockRay { get; set; }
+		internal static RayLockingDelegate unlockRay { get; set; }
+
 		/// <summary>
 		/// Lock the default ray's show/hide state.
-		/// Transform = Ray origin
-		/// object = The object performing the lock is passed in and must be used for unlocking
 		/// </summary>
-		Func<Transform, object, bool> lockRay { set; }
+		/// <param name="rayOrigin">The ray to lock</param>
+		/// <param name="obj">The object performing the lock is passed in and must be used for unlocking</param>
+		public static void LockRay(this IUsesRayLocking customRay, Transform rayOrigin, object obj)
+		{
+			if (lockRay != null)
+				lockRay(rayOrigin, obj);
+		}
 
 		/// <summary>
 		/// Unlock the default ray's show/hide state.
-		/// Transform = Ray origin
-		/// object = The object performing the unlock must be passed in and match the one that locked it or null to override
 		/// </summary>
-		Func<Transform, object, bool> unlockRay { set; }
+		/// <param name="rayOrigin">The ray to unlock</param>
+		/// <param name="obj">The object performing the unlock must be passed in and match the one that locked it or null to override</param>
+		public static void UnlockRay(this IUsesRayLocking customRay, Transform rayOrigin, object obj)
+		{
+			if (unlockRay != null)
+				unlockRay(rayOrigin, obj);
+		}
 	}
+
 }
 #endif
