+++ Scripts/Utilities/ObjectUtils.cs
@@ -4,6 +4,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using UnityEditor.Experimental.EditorVR.Core;
 using UnityEngine;
 using UnityObject = UnityEngine.Object;
 
@@ -61,7 +62,7 @@ namespace UnityEditor.Experimental.EditorVR.Utilities
 		public static GameObject CreateEmptyGameObject(String name = null, Transform parent = null)
 		{
 			GameObject empty = null;
-			if (string.IsNullOrEmpty(name))
+			if (String.IsNullOrEmpty(name))
 				name = "Empty";
 
 #if UNITY_EDITORVR
@@ -222,6 +223,13 @@ namespace UnityEditor.Experimental.EditorVR.Utilities
 				return GetAssignableTypes(type);
 
 			return Enumerable.Empty<Type>();
+		}
+
+		public static bool IsSameAssembly<T>(object obj)
+		{
+			// Until we move EditorVR into it's own assembly, this is a way to enforce 'internal' on interfaces
+			var objType = obj.GetType();
+			return objType.Assembly == typeof(T).Assembly;
 		}
 
 		public static void Destroy(UnityObject o, float t = 0f)
